Change everything to the page
