+++ src/components/ButtonComponent/index.js
@@ -1,10 +1,10 @@
 import React from 'react';
-import {ReactComponent as ButtonPlus} from '../../assets/images/buttonplus.svg';
+import { ReactComponent as ButtonPlus } from '../../assets/images/buttonplus.svg';
 import './ButtonComponent.css';
 
-const CreateButton =(props) =>( 
+const CreateButton = (props) => (
   <button className="RoundAddButton" onClick={props.onClick}>
-      < ButtonPlus/>
+    <ButtonPlus />
   </button>
 );
