+++ src/components/Footer/ClustersList/ClustersList.css
