+++ src/components/PrimaryButton/index.js
@@ -3,12 +3,14 @@ import './PrimaryButton.css';
 
 const PrimaryButton = (props) => {
   const { label } = props;
-  return <button
-    className="Primary-Btn uppercase"
-    onClick={props.onClick}
-  >
-    {label}
-  </button>;
+  return (
+    <button
+      className="Primary-Btn uppercase"
+      onClick={props.onClick}
+    >
+      {label}
+    </button>
+  );
 };
 
 export default PrimaryButton;
