+++ src/components/ResourceCard/index.js
@@ -4,7 +4,7 @@ import './ResourceCard.css';
 const ResourceCard = (props) => (
     <div className="Card">
       <div className="CardHeader">{props.title}</div>
-      <div className="ResourceDigit">{props.resourceCount}</div>
+      <div className="ResourceDigit">{props.count}</div>
     </div>
   )
